extract the test which needs Docker apart as Testcontainers annotation is supported only at class level
